RTE-941: Move common functions to parent image builder.
